Add files via upload
it is played with two player and two 6 sided dices. in the first Turn 6+6, 5+5,3+3 is a winner. 4+4, 2+2, 1+1 is a loser, Also in the first turn if the total is 7 you are a winner. 
in the second and consecutive   turns you try to get the total of the previous turn. if you get you are a winner. İf your total is 7 you lose.